requirements.txt
absl-py
aiohttp
alembic
anyio
argon2-cffi
asttokens
async-timeout
attrs
autopage
babel
backcall
backports.functools-lru-cache
black
bleach
blinker
bottle
bottleneck
cachetools
certifi
cffi
chardet
charset-normalizer
click
cliff
cmaes
cmd2
colorlog
cryptography
cycler
cython
debugpy
decorator
defusedxml
entrypoints
executing
filelock
flit-core
fonttools
fsspec
future
geotorch
google-auth
google-auth-oauthlib
greenlet
grpcio
h5py
huggingface-hub
idna
importlib-metadata
importlib-resources
ipykernel
ipython
ipython-genutils
jedi
jinja2
joblib
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
kiwisolver
markdown
markupsafe
matplotlib
matplotlib-inline
mistune
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numexpr
oauthlib
optuna-dashboard
packaging
pandocfilters
parso
pathspec
pbr
pexpect
pickleshare
pillow
pip
platformdirs
prettytable
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pycparser
pydeprecate
pydocstyle
pygments
pyjwt
pyopenssl
pyparsing
pyperclip
pyrsistent
pysocks
python-dateutil
pytorch-lightning
pyu2f
pyyaml
pyzmq
requests
requests-oauthlib
rsa
sacremoses
scikit-learn
send2trash
setuptools
six
sniffio
snowballstemmer
sqlalchemy
stack-data
stevedore
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
terminado
testpath
threadpoolctl
tornado
tqdm
traitlets
transformers
typed-ast
typing-extensions
typing_extensions
wcwidth
webencodings
websocket-client
werkzeug
wheel
yarl
zipp

# pip-only extras
appdirs
astor
autopep8
databind
databind-core
databind-json
deprecated
docker-pycreds
docspec
docspec-python
docstr-md
docstring-parser
ecos
flake8
gitdb
gitpython
iniconfig
mccabe
metric-learn
nr-util
numpy
optuna
pandas
pathlib
pathtools
patsy
pluggy
psutil
py
pycodestyle
pydoc-markdown
pyflakes
pytest
pytz
quantificationlib
scipy
sentry-sdk
setproctitle
smmap
statsmodels
toml
tomli-w
torch-tb-profiler
tzdata
urllib3
wandb
watchdog
wrapt
yapf
